Adds dTDA
