+++ momentGW/__init__.py
@@ -49,4 +49,6 @@ except ImportError:
 from momentGW.evgw import evGW
 from momentGW.gw import GW
 from momentGW.qsgw import qsGW
+from momentGW.rpa import RPA
 from momentGW.scgw import scGW
+from momentGW.tda import TDA
